+++ src/main/java/com/leets/commitatobe/domain/login/dto/GitHubDto.java
@@ -1,0 +1,6 @@
+package com.leets.commitatobe.domain.login.dto;
+
+public record GitHubDto(
+	String userId
+) {
+}
