--- src/main/java/com/leets/commitatobe/domain/login/presentation/dto/GitHubDto.java
@@ -1,6 +1,0 @@
-package com.leets.commitatobe.domain.login.presentation.dto;
-
-
-public record GitHubDto(
-    String userId
-) {}
