+++ src/main/java/com/leets/commitatobe/domain/login/service/LoginQueryService.java
@@ -1,11 +1,11 @@
-package com.leets.commitatobe.domain.login.usecase;
+package com.leets.commitatobe.domain.login.service;
 
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import com.leets.commitatobe.domain.login.domain.CustomUserDetails;
-import com.leets.commitatobe.domain.login.presentation.dto.GitHubDto;
+import com.leets.commitatobe.domain.login.dto.GitHubDto;
 import com.leets.commitatobe.global.exception.ApiException;
 import com.leets.commitatobe.global.response.code.status.ErrorStatus;
 
@@ -16,9 +16,8 @@ import lombok.extern.slf4j.Slf4j;
 @Service
 @RequiredArgsConstructor
 @Slf4j
-public class LoginQueryServiceImpl implements LoginQueryService {
+public class LoginQueryService {
 
-	@Override
 	public GitHubDto getGitHubUser(HttpServletRequest request) {
 		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 		if (authentication == null || !(authentication.getPrincipal() instanceof CustomUserDetails)) {
@@ -29,7 +28,6 @@ public class LoginQueryServiceImpl implements LoginQueryService {
 		return userDetails.getGitHubDto();
 	}
 
-	@Override
 	public String getGitHubId(HttpServletRequest request) {
 		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 		if (authentication == null || !(authentication.getPrincipal() instanceof CustomUserDetails)) {
